+++ front/scripts/auth/common/AuthTracker.ts
@@ -1,11 +1,19 @@
 class AuthTracker {
 	gaCategory: string;
 
-	constructor (gaCategory: string) {
+	constructor (gaCategory: string, pageType: string) {
 		this.gaCategory = gaCategory;
+		M.setTrackContext({
+			a: pageType,
+			n: -1
+		});
 	}
 
-	public trackClick (element: HTMLElement, label: string, action = Mercury.Utils.trackActions.click): void {
+	private setGaCategory (page: string): string {
+		return 'user-' + page + '-' + pageParams.viewType + (pageParams.isModal ? '-modal' : '');
+	}
+
+	public trackClick (element: HTMLElement, label: string, action = M.trackActions.click): void {
 		if (!element) {
 			return;
 		}
@@ -15,10 +23,8 @@ class AuthTracker {
 		}.bind(this))
 	}
 
-	public trackPageView (pageType: string) {
-		if (pageType) {
-			this.track(pageType, M.trackActions.impression);
-		}
+	public trackPageView () {
+		M.trackPageView(null);
 	}
 
 	public trackSubmit (form: HTMLFormElement, label: string): void {
@@ -32,11 +38,18 @@ class AuthTracker {
 	}
 
 	public track (label: string, action: string) {
-		M.track({
-			trackingMethod: 'both',
-			action: action,
-			category: this.gaCategory,
-			label: label
-		});
+		var trackOptions: TrackingParams = {
+				trackingMethod: 'both',
+				action: action,
+				category: this.gaCategory,
+				label: label
+			},
+			sourceUrl = M.getQueryParam('redirect');
+
+		if (sourceUrl) {
+			trackOptions.sourceUrl = sourceUrl;
+		}
+
+		M.track(trackOptions);
 	}
 }