+++ front/scripts/auth/common/Login.ts
@@ -24,6 +24,7 @@ class Login {
 	usernameInput: HTMLInputElement;
 	passwordInput: HTMLInputElement;
 	urlHelper: UrlHelper;
+	tracker: AuthTracker;
 
 	constructor (form: Element) {
 		var elements: FormElements;
@@ -38,6 +39,7 @@ class Login {
 			this.redirect = params['redirect'];
 		}
 		this.redirect = this.redirect || '/';
+		this.tracker = new AuthTracker('user-login-mobile', '/signin');
 	}
 
 	public onSubmit (event: Event): void {
@@ -61,10 +63,10 @@ class Login {
 			enableSubmitButton();
 
 			if (xhr.status === HttpCodes.UNAUTHORIZED) {
-				this.track('login-credentials-error', Mercury.Utils.trackActions.error);
+				this.tracker.track('login-credentials-error', M.trackActions.error);
 				return this.displayError('errors.wrong-credentials');
 			} else if (xhr.status !== HttpCodes.OK) {
-				this.track('login-server-error', Mercury.Utils.trackActions.error);
+				this.tracker.track('login-server-error', M.trackActions.error);
 				return this.displayError('errors.server-error');
 			}
 
@@ -72,7 +74,7 @@ class Login {
 
 			if (response.error) {
 				// Helios may return an error even if the request returns a 200
-				this.track('login-credentials-error', Mercury.Utils.trackActions.error);
+				this.tracker.track('login-credentials-error', M.trackActions.error);
 				this.displayError('errors.wrong-credentials');
 			} else {
 				this.onLoginSuccess(response);
@@ -82,7 +84,7 @@ class Login {
 		xhr.onerror = (): void => {
 			enableSubmitButton();
 
-			this.track('login-server-error', Mercury.Utils.trackActions.error);
+			this.tracker.track('login-server-error', M.trackActions.error);
 			this.displayError('errors.server-error');
 		};
 
@@ -93,12 +95,20 @@ class Login {
 	}
 
 	public onLoginSuccess(loginResponse: LoginResponse): void {
-		this.track('login-success', Mercury.Utils.trackActions.submit);
-		window.location.href = this.redirect;
+		this.tracker.track('login-success', M.trackActions.submit);
+		Utils.loadUrl(this.redirect);
 	}
 
 	public watch(): void {
 		this.form.addEventListener('submit', this.onSubmit.bind(this));
+
+		// TODO remove when SOC-719 is ready
+		if (pageParams.isModal) {
+			this.form.querySelector('.forgotten-password').addEventListener('click', function(event) {
+				Utils.loadUrl((<HTMLLinkElement> event.target).href);
+				event.preventDefault();
+			});
+		}
 	}
 
 	private getCredentials (): LoginCredentials {
@@ -120,15 +130,6 @@ class Login {
 		if (errorNode) {
 			errorNode.parentNode.removeChild(errorNode);
 		}
-	}
-
-	private track (label: string, action: string): void {
-		M.track({
-			trackingMethod: 'both',
-			action: action,
-			category: 'user-login-' + pageParams.viewType,
-			label: label
-		});
 	}
 }