+++ front/scripts/auth/facebook/FacebookRegistration.ts
@@ -28,6 +28,8 @@ class FacebookRegistration {
 	marketingOptIn: MarketingOptIn;
 	formErrors: FormErrors;
 	termsOfUse: TermsOfUse;
+	tracker: AuthTracker;
+	utils: Utils;
 
 	constructor (form: HTMLFormElement) {
 		new FacebookSDK(this.init.bind(this));
@@ -46,6 +48,7 @@ class FacebookRegistration {
 		this.formErrors = new FormErrors(this.form, 'fbRegistrationValidationErrors');
 
 		this.form.addEventListener('submit', this.onSubmit.bind(this));
+		this.tracker = new AuthTracker('user-signup-mobile', 'signup');
 	}
 
 	public init (): void {
@@ -98,8 +101,8 @@ class FacebookRegistration {
 			var status: number = (<XMLHttpRequest> e.target).status;
 
 			if (status === HttpCodes.OK) {
-				this.track('facebook-signup-join-wikia-success', Mercury.Utils.trackActions.success);
-				window.location.href = this.redirect;
+				this.tracker.track('facebook-signup-join-wikia-success', Mercury.Utils.trackActions.success);
+				Utils.loadUrl(this.redirect);
 			} else if (status === HttpCodes.BAD_REQUEST) {
 				this.formErrors.displayGeneralError();
 			} else {
@@ -109,7 +112,7 @@ class FacebookRegistration {
 
 		facebookTokenXhr.onerror = (e: Event): void => {
 			this.formErrors.displayGeneralError();
-			this.track('facebook-signup-join-wikia-error', Mercury.Utils.trackActions.error);
+			this.tracker.track('facebook-signup-join-wikia-error', Mercury.Utils.trackActions.error);
 		};
 
 		facebookTokenXhr.open('POST', heliosTokenUrl, true);
@@ -141,21 +144,12 @@ class FacebookRegistration {
 
 		facebookRegistrationXhr.onerror = (e: Event) => {
 			this.formErrors.displayGeneralError();
-			this.track('facebook-signup-join-wikia-error', Mercury.Utils.trackActions.error);
+			this.tracker.track('facebook-signup-join-wikia-error', Mercury.Utils.trackActions.error);
 		};
 
 		facebookRegistrationXhr.open('POST', url, true);
 		facebookRegistrationXhr.withCredentials = true;
 		facebookRegistrationXhr.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
 		facebookRegistrationXhr.send(this.urlHelper.urlEncode(data));
-	}
-
-	private track (label: string, action: string): void {
-		M.track({
-			trackingMethod: 'both',
-			action: action,
-			category: 'user-signup-mobile',
-			label: label
-		});
 	}
 }