+++ front/scripts/main/routes/CuratedContentEditorRoute.ts
@@ -118,7 +118,16 @@ App.CuratedContentEditorRoute = Em.Route.extend(
 			openMainPage(dataSaved: boolean = false): void {
 				var ponto = window.Ponto;
 
+				this.set('publish', !!dataSaved);
+
 				if (ponto && typeof ponto.invoke === 'function') {
+
+					if (App.CuratedContentEditorModel.isDirty &&
+						!this.get('publish') &&
+						!confirm(i18n.t('app.curated-content-editor-exit-prompt'))
+					) {
+						return;
+					}
 					ponto.invoke(
 						// AMD module name in app
 						'curatedContentTool.pontoBridge',
@@ -141,7 +150,7 @@ App.CuratedContentEditorRoute = Em.Route.extend(
 						},
 						true
 					);
-				} else {
+			} else {
 					this.transitionTo('mainPage');
 				}
 			},
@@ -171,11 +180,23 @@ App.CuratedContentEditorRoute = Em.Route.extend(
 			 * @returns {boolean}
 			 */
 			willTransition(transition: EmberStates.Transition): boolean {
-				if (transition.targetName.indexOf('curatedContentEditor') < 0) {
+				var isStayingOnEditor: boolean = transition.targetName.indexOf('curatedContentEditor') > -1;
+
+				if (
+					App.CuratedContentEditorModel.isDirty &&
+					!isStayingOnEditor &&
+					!this.get('publish') &&
+					!confirm(i18n.t('app.curated-content-editor-exit-prompt'))
+				) {
+					transition.abort();
+				}
+
+				if (!isStayingOnEditor) {
 					transition.then(() => {
 						this.controllerFor('application').set('fullPage', false);
 					});
 				}
+
 				return true;
 			},