+++ server/facets/auth/authView.ts
@@ -19,9 +19,8 @@ module authView {
 		language: string;
 		exitTo: string;
 		mainPage: string;
-		standalonePage: boolean;
 		optimizelyScript: string;
-		pageParams: PageParams;
+		pageParams: any;
 		hideHeader?: boolean;
 		hideFooter?: boolean;
 		footerHref?: string;
@@ -29,8 +28,9 @@ module authView {
 		footerCalloutLink?: string;
 		headerText?: string;
 		bodyClasses?: string;
+		pageType?: string;
+		standalone?: boolean;
 		trackingConfig?: any;
-		isModal?: boolean;
 	}
 
 	export function view (template: string, context: AuthViewContext, request: Hapi.Request, reply: any): Hapi.Response {
@@ -90,13 +90,13 @@ module authView {
 			canonicalUrl: this.getCanonicalUrl(request),
 			exitTo: this.getRedirectUrl(request),
 			mainPage: 'http://www.wikia.com',
-			isModal: isModal,
 			language: request.server.methods.i18n.getInstance().lng(),
 			trackingConfig: localSettings.tracking,
 			optimizelyScript: localSettings.optimizely.scriptPath +
-				localSettings.optimizely.account + '.js',
+			localSettings.optimizely.account + '.js',
 			standalonePage: (viewType === authView.VIEW_TYPE_DESKTOP && !isModal),
 			pageParams: {
+				isModal: isModal,
 				viewType: viewType
 			}
 		};